Combining satellite & sbdb renderables under a common base class
