+++ modules/space/rendering/renderablesatellites.h
@@ -25,6 +25,7 @@
 #ifndef __OPENSPACE_MODULE_SPACE___RENDERABLESATELLITES___H__
 #define __OPENSPACE_MODULE_SPACE___RENDERABLESATELLITES___H__
 
+#include <modules/space/rendering/renderableorbitalkepler.h>
 #include <openspace/rendering/renderable.h>
 
 #include <modules/base/rendering/renderabletrail.h>
@@ -37,98 +38,13 @@
 
 namespace openspace {
 
-const std::vector<int> LeapYears = {
-    1956, 1960, 1964, 1968, 1972, 1976, 1980, 1984, 1988, 1992, 1996,
-    2000, 2004, 2008, 2012, 2016, 2020, 2024, 2028, 2032, 2036, 2040,
-    2044, 2048, 2052, 2056
-};
-int countDays(int year);
-int countLeapSeconds(int year, int dayOfYear);
-double calculateSemiMajorAxis(double meanMotion);
-double epochFromSubstring(const std::string& epochString);
-
-class RenderableSatellites : public Renderable {
+class RenderableSatellites : private RenderableOrbitalKepler {
 public:
     RenderableSatellites(const ghoul::Dictionary& dictionary);
-
-    void initializeGL() override;
-    void deinitializeGL() override;
-
-    bool isReady() const override;
-    void render(const RenderData& data, RendererTasks& rendererTask) override;
-
     static documentation::Documentation Documentation();
-    /**
-        * Reads the provided TLE file and calls the KeplerTranslation::setKeplerElments
-        * method with the correct values. If \p filename is a valid TLE file but contains
-        * disallowed values (see KeplerTranslation::setKeplerElements), a
-        * KeplerTranslation::RangeError is thrown.
-        *
-        * \param filename The path to the file that contains the TLE file.
-        *
-        * \throw ghoul::RuntimeError if the TLE file does not exist or there is a 
-        *        problem with its format.
-        * \pre The \p filename must exist
-        */
-    void readTLEFile(const std::string& filename);
 
 private:
-    struct Vertex {
-        glm::vec3 position = glm::vec3(0.f);
-        glm::vec3 color = glm::vec3(0.f);
-        glm::vec2 texcoord = glm::vec2(0.f);
-    };
 
-    struct KeplerParameters {
-        double inclination = 0.0;
-        double semiMajorAxis = 0.0;
-        double ascendingNode = 0.0;
-        double eccentricity = 0.0;
-        double argumentOfPeriapsis = 0.0;
-        double meanAnomaly = 0.0;
-        double meanMotion = 0.0;
-        double epoch = 0.0;
-        double period = 0.0;
-    };
-
-    /// The layout of the VBOs
-    struct TrailVBOLayout {
-        float x = 0.f;
-        float y = 0.f;
-        float z = 0.f;
-        float time = 0.f;
-        double epoch = 0.0;
-        double period = 0.0;
-    };
-
-    KeplerTranslation _keplerTranslator;
-    std::vector<KeplerParameters> _TLEData;
-
-    /// The backend storage for the vertex buffer object containing all points for this
-    /// trail.
-    std::vector<TrailVBOLayout> _vertexBufferData;
-
-    GLuint _vertexArray;
-    GLuint _vertexBuffer;
-    GLuint _indexBuffer;
-
-    //GLuint _vaoTest; // vertexArrayObject
-    //GLuint _vboTest; // vertextBufferObject
-    //GLuint _eboTest; // elementBufferObject/ indexBufferObject       
-
-    void updateBuffers();
-
-    ghoul::opengl::ProgramObject* _programObject;
-
-    properties::StringProperty _path;
-    properties::UIntProperty _nSegments;
-
-    RenderableTrail::Appearance _appearance;
-
-    glm::vec3 _position = glm::vec3(0.f);
-
-    UniformCache(modelView, projection, lineFade, inGameTime, color, opacity,
-        numberOfSegments) _uniformCache;
 };
 
 } // namespace openspace
